Create a navbar
This commit creates a navbar which renders the brand, links to view
Recipes and Shopping List. These links do not navigate yet. A button
labeled "Manage" is also rendered. There is also markup in place for
the Manage button to toggle a drop down list. The drop down behavior
is not implemented. There is no way to view the items in the drop down
list. The dropdown list contains markup to display "Save Data" and
"Fetch Data" menu options. Markup is in place to use Bootstrap's
responsive collapse navbar component. However, the behavior is not yet
implemented.