Render recipe data for each recipe in list
Modify RecipeListComponent to render each recipe in an array of recipes.
This change adds a recipes property to the component class. A Recipe
model is created for representing recipe data. There are 3 public
properties on the Recipe class:
  - name: string - title for the recipe
  - description: string
  - imagePath: string - url of image associated to the recipe
There is a constructor which accepts three arguments: name, description,
and imagePath. The arguments are used to set the initial value for the
public properties.