Create initial template for recipe details
Modify the template for RecipeDetailComponent. The recipe detail renders
the template with place holder data including the Recipe Name,
Description, Ingredients, and Manage Recipe button. The button does not
have any behavior yet. The markup for the menu that will be opened by
the "Manage Recipe" button is present, but not visible.

Use a templateUrl instead of an inline template for
RecipeDetailComponent.