Create Ingredient model
Create Ingredient model which is a class for transferring data for an
ingredient. The model has two properties as follows:
  - name: string
  - amount: number - units of the ingredient

The constructor accepts a name and amount parameter to set the initial
values of these properties.

Ingredient has been placed in app/shared because this model will be
used in both shopping-list and recipes views.