Refactor recipe list
Refactor recipe list to project recipe list items into recipe list.
This changes the the way recipe list works. The recipe-list component
is a wrapper to format the list. The state of the recipe list is now
tracked by the recipes component. This change has been implemented to
allow for a cleaner binding to the recipe item select event in the
recipes component instead of passing the event up through the
recipe-list component.