User can add items to shopping list
Shopping list edit component emits addIngredient event when the add
button is clicked. The even emits with an Ingredient for the name and
amount from the form inputs.

Shoping list component listens for the addIngredient event and updates
the displayed list of ingredients with the new Ingredient.

FormsModule is imported by App component.