Implement RecipeService
Refactor components for Recipe list and detail views to implement
RecipeService provider to handle flow of recipe data between views.
The RecipeComponent stores state for the currently selected receipe.

RecipeService#getRecipes returns an array of Recipes.
ReciperService#recipeSelect is an EventEmmiter which emits a
Recipe.

recipe-list-item component calls recipeSelect.emit to notify
subscribers of the event of the recipe clicked on by the user.

recipe component subscribes to recipeSelect and update the
selectedRecipe when a recipe is selected.