Define ShoppingListService and implement
Create ShoppingListService. Implement the service to manage
Ingredients in the shopping list.

ShoppingListService#getIngredients returns an array of Ingredients
which make up the shopping list.

ShoppingListService#ingredientChange is an EventEmitter which emits
an array for the current list of Ingredients.

ShoppingListService#addIngredient(ingredient: Ingredient) adds
an Ingredient to the shopping list and emits ingredientChange.

Shopping-list-component initializes with the shopping list
Ingredients from the service and subscribes to the ingredientChange
event to update the shopping list items that are rendered.

Shopping-list-edit component calls addIngredient when user adds
a new ingredient.