User can add recipe ingredients to shopping list
All ingredients for a recipe are added to the shopping list when the
Manage Recipe>To Shopping List control is clicked. This is done
by calling addIngredientsToShopping list method on the detail component,
which adds the ingredients using the ShoppingListService.

Add ingredients: Ingredient[] to recipe.model.

Recipe Detail component depends on ShoppingService instance provided
by Angular.