Add features to shopping list service
Add editItem: Subject property to shopping list service. This
will emit a number which corresponds to the index in #ingredients
for the ingredit opened for editing.

Add getIngredientAt, changeIngredientAt, and removeIngredientAt
methods to shopping-list service.

getIngredientAt(index: number): Ingredient returns the ingredient
at the given index in #ingredients.

removeIngredientAt(index: number): Ingredient removes the element
for the given index in #ingredients and emits ingredientChange
with #ingredients after the change has taken place.

changeIngredientAt(ingredient: Ingredient, index: number): self
replaces the element for the given index in #ingredients. This
method may be chained. (Returns instance of itself.)