Can edit and delete shopping list items
Clicking an item in the shopping list opens the item for editing
in the shopping-list-item-edit form and toggles the editMode
property on the edit component.

The "Add" button text is changed to "Update" and clicking the
button updates the ingredient in the ingredients service from the
values of the edit form.

Clicking on "delete" button will remove the ingredient currently
loaded in the edit form from the ingredients service.

Clicking on "clear" resets the form values.