Submit recipe edit form to create or modify recipe
Add onSubmit method to edit-recipe component. Calling onSubmit will
create a new Recipe from the edit form and update the recipe if the
url path is './:id/edit'. Otherwise, a new recipe is added to the
recipe list by called the addRecipe method of the recipe service. The
app navigates to the detail view for the edited recipe.

The onSubmit method is registered to handle the submit event on the
recipe edit form.

Recipe list subscribes to recipe service recipesChange subject and
syncs the displayed recipe list when data is emitted.