Data storage service can save and fetch recipes
Add two methods to data storage service.

save(): Observable sends a request to update the recipes in
the database. An observable which emits an array of the sent recipe data
is returned.

fetch(): Observable requests recipe data from the database and updates
the recipes stored by the recipe service. An observable which emits an
array of recipes from the database response is returned.