Create signup component and routing
Add '/signup' route to app-routing which displays the component.
The component renders a form with controls to get user email and
password. The onSignup method of the component is called when
the form is submitted.

The email and password are logged to the console when the form
is submitted. The onsignUp method should be modified in the
future to use the auth service to create a new user.

Add link to top level navigation bar. Click on "Rigister" navigates
to "/signup".