Create AuthService module
Create AuthService to handle user authentication with the
firebase sdk.

Create signupUser method which accepts an email: string, and
password: string and creates a new user account with the firebase
sdk. The error is logged in the console if the create user promise
from firebase rejects. Call authService.signupUser(email, password)
to create a new user account.

Service depends on firebase/app and firebase/auth.