Create the Signin component and routing
This commit create a signin component which is responsible for
rendering a form to get an existing user's email and password.

The onSignin() method is called when the form is submitted. This
method gets the user input and logs it in the console. This method
should be modified to use the auth service to authenticate the
user.