AuthService#signinUser() authenticates user
This commmit create the signinUser method of the auth service.
Call `authService.signinUser(email, password)` to use the firebase
sdk to authenticate the user given email: string, and password:
string.

The user object resolved by firebase is logged in the console.

The error object is logged in the console if the firebase promise
rejects.