AuthService provides a token
This commit adds getToken() method to the auth service. Whenever
signupUser or singinUser is called, a reference to the token is
created as a private property of the auth service. Calling getToken
uses the firebase sdk to store the token, update the token if it
has expired, and return the token.