fix: AuthService#getToken() returns current token
This commit fixes the issue where getToken() would sometimes
return an expired token by changeing the function to return a
promise for the token. This change breaks the signature of the
method because it returns a Promise instead of a string.

Modify data storage service methods to use getToken() asynchronously.