refactor: Use interceptor for auth http params
This commit creates the AuthInterceptor class which is implemented as
a provider in the CoreModule. AuthInterceptor is reponsible for
intercepting http requests made with httpClient and adding
authentications before the request is sent.

Modify DataStorage service so that it no longer depends on the
AuthService. All requests made by this service are implicitly
intercepted and modified to include an 'auth' query param by the
AuthInterceptor.