feat: Create shopping-list store
This commit adds @ngrx/store to the project dependencies and creates
an intial store with files for exporting reducers and actions to manage
the state of the shoppingList.

State's interface:
  {
    shoppingList: { ingredients: Ingredient[] }
  }

Create ADD_INGREDIENT action type and AddIngredient action. The action
accepts an Ingredient as the payload. The reducer adds the payload to
the state.shoppingList.ingredients when the action is dispatched.