feat: Create auth store
The commit create reducers, actions, and effects to manage the
authentication state of the app and side effects of changes in the
state.

See a list of action types exported as constants listed near the top
of auth.actions.

The following actions will have side effects for managing user auth
with firebase:
  - TRY_SIGNUP: try to create a new user with firebase.auth() and get
    a token

  - TRY_SIGNIN: try to signin user with firebase.auth() and get token

  - LOGOUT: sign out of firebase.auth()